fix: need casts in tests

+++ tests/jax/test_jax_utilities.py
@@ -11,7 +11,7 @@ def test_jax_utilities_horner_dtype():
     truth = coef[0] + coef[1] * x + coef[2] * x**2 + coef[3] * x**3 + coef[4] * x**4
 
     result = jax_galsim.utilities.horner(x, coef, dtype=int)
-    np.testing.assert_almost_equal(result, truth)
+    np.testing.assert_almost_equal(result, truth.astype(int))
 
     result = jax_galsim.utilities.horner(x, coef, dtype=float)
     np.testing.assert_almost_equal(result, truth)
@@ -21,19 +21,19 @@ def test_jax_utilities_horner_dtype():
     np.testing.assert_almost_equal(result.imag, np.zeros_like(truth))
 
     result = jax_galsim.utilities.horner(x, coef, dtype=jnp.int32)
-    np.testing.assert_almost_equal(result, truth)
+    np.testing.assert_almost_equal(result, truth.astype(np.int32))
 
     result = jax_galsim.utilities.horner(x, coef, dtype=jnp.int64)
-    np.testing.assert_almost_equal(result, truth)
+    np.testing.assert_almost_equal(result, truth.astype(np.int64))
 
     result = jax_galsim.utilities.horner(x, coef, dtype=jnp.float32)
-    np.testing.assert_almost_equal(result, truth)
+    np.testing.assert_almost_equal(result, truth.astype(np.float32))
 
     result = jax_galsim.utilities.horner(x, coef, dtype=jnp.float64)
     np.testing.assert_almost_equal(result, truth)
 
     result = jax_galsim.utilities.horner(x, coef, dtype=jnp.complex64)
-    np.testing.assert_almost_equal(result.real, truth)
+    np.testing.assert_almost_equal(result.real, truth.astype(np.complex64))
     np.testing.assert_almost_equal(result.imag, np.zeros_like(truth))
 
     result = jax_galsim.utilities.horner(x, coef, dtype=jnp.complex128)
